pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "canary-release-env"
version = "1.0.0"
description = "OpenEnv environment for ML model canary release management"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "openai>=2.30.0",
]

[project.scripts]
server = "canary_release_env.server.app:main"
review-artifacts = "canary_release_env.generate_review_artifacts:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel.force-include]
"__init__.py" = "canary_release_env/__init__.py"
"models.py" = "canary_release_env/models.py"
"client.py" = "canary_release_env/client.py"
"baseline.py" = "canary_release_env/baseline.py"
"inference.py" = "canary_release_env/inference.py"
"generate_review_artifacts.py" = "canary_release_env/generate_review_artifacts.py"
"server" = "canary_release_env/server"
"openenv.yaml" = "canary_release_env/openenv.yaml"

[tool.pytest.ini_options]
asyncio_mode = "auto"
